refactor: API/Action 레이어 7개 -> 4개 파일로 구조 단순화
- lib/api/backend.ts [NEW]: fetchBackend + refreshTokens + getBackendBaseUrl 통합
  (구 client.ts + auth-service.ts + bff-utils.ts BFF 로직 흡수)
- lib/utils/auth.ts [NEW]: 인증 상수 + 클라이언트 유틸 + 쿠키 프록시 유틸 통합
  (구 auth-utils.ts + bff-utils.ts 쿠키 파트 흡수)
- lib/api/client.ts, lib/services/auth-service.ts, lib/utils/bff-utils.ts,
  lib/utils/auth-utils.ts 삭제 (-4개 파일)
- lib/services/ 폴더 제거
- middleware.ts: BffInfo 타입/authService/getBffInfo 제거, isLocalRequest() 헬퍼 인라인화
- 모든 참조 파일 import 경로 일괄 업데이트
