+++ apps/web/app/api/auth/[provider]/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 
-import { fetchBackend } from '@/lib/api/client';
-import { getBffInfo, proxyCookies } from '@/lib/utils/bff-utils';
+import { fetchBackend, getBackendBaseUrl } from '@/lib/api/backend';
+import { proxyCookies } from '@/lib/utils/auth';
 
 /**
  * BFF (Backend For Frontend) OAuth Authorization Redirect Handler
@@ -18,8 +18,11 @@ export async function GET(
     return NextResponse.json({ error: 'Invalid auth provider' }, { status: 400 });
   }
 
-  const bffInfo = getBffInfo(request);
-  const backendUrl = `${bffInfo.backendUrl}/oauth2/authorization/${provider}`;
+  const backendUrl = `${getBackendBaseUrl()}/oauth2/authorization/${provider}`;
+  const isLocal =
+    request.nextUrl.hostname === 'localhost' ||
+    request.nextUrl.hostname === '127.0.0.1' ||
+    request.nextUrl.protocol === 'http:';
 
   try {
     const response = await fetchBackend(backendUrl, {
@@ -33,10 +36,7 @@ export async function GET(
 
     if (redirectUrl) {
       const res = NextResponse.redirect(redirectUrl);
-
-      // 백엔드 쿠키들을 브라우저로 전달
-      proxyCookies(response, res, bffInfo.isLocal);
-
+      proxyCookies(response, res, isLocal);
       return res;
     }
