--- apps/web/lib/services/auth-service.ts
@@ -1,98 +1,0 @@
-/**
- * Auth Service — 백엔드 인증 API 통신 서비스
- *
- * [Edge Runtime 호환]
- * - 순수 fetch API만 사용
- * - Node.js 전용 모듈(fs, http, crypto 등) 사용 금지
- *
- * [사용처]
- * - middleware.ts (Edge Runtime)
- * - app/api/auth/refresh/route.ts (Node.js Runtime)
- */
-
-import { NextRequest } from 'next/server';
-
-import { fetchBackend } from '@/lib/api/client';
-
-export interface RefreshResult {
-  accessToken: string;
-  /** 백엔드가 발급한 모든 Set-Cookie 헤더 목록 */
-  setCookies: string[];
-  message: string;
-}
-
-/**
- * 백엔드 /api/auth/refresh 응답 Body 구조
- */
-interface BackendRefreshResponse {
-  status: string;
-  message: string;
-  data: {
-    accessToken: string;
-  } | null;
-}
-
-export const authService = {
-  /**
-   * refreshToken을 사용하여 새로운 accessToken을 발급받습니다.
-   *
-   * @param bffRequest - (옵션) Middleware에서 호출 시 현재 요청 객체
-   * @throws Error - 네트워크 오류, 백엔드 오류 등
-   */
-  async refreshAccessToken(bffRequest?: NextRequest): Promise<RefreshResult> {
-    const refreshUrl = `/api/auth/refresh`;
-
-    let response: Response;
-    try {
-      response = await fetchBackend(refreshUrl, {
-        method: 'POST',
-        bffRequest,
-        signal: AbortSignal.timeout(5000),
-      });
-    } catch (networkError) {
-      throw new Error(`[AuthService] 네트워크 오류: ${String(networkError)}`);
-    }
-
-    if (!response.ok) {
-      let errorDetail = '';
-      try {
-        errorDetail = await response.text();
-      } catch {
-        errorDetail = '(응답 body 없음)';
-      }
-
-      // 구체적인 에러 로깅
-      console.error(
-        `[AuthService] 토큰 재발급 실패 | Status: ${response.status} | URL: ${refreshUrl}`,
-        {
-          statusText: response.statusText,
-          body: errorDetail,
-        },
-      );
-
-      throw new Error(`백엔드 토큰 재발급 실패 (HTTP ${response.status}): ${errorDetail}`);
-    }
-
-    let body: BackendRefreshResponse;
-    try {
-      body = await response.json();
-    } catch (e) {
-      console.error('[AuthService] JSON 파싱 에러:', e);
-      throw new Error('백엔드 응답 데이터 형식이 올바르지 않습니다.');
-    }
-
-    if (!body.data?.accessToken) {
-      console.error('[AuthService] 액세스 토큰 누락:', body);
-      throw new Error(`토큰 발급 실패: ${body.message || '데이터 구조 오류'}`);
-    }
-
-    // 백엔드에서 내려온 모든 쿠키(토큰 회전 포함)를 캡처
-    const setCookies = response.headers.getSetCookie();
-
-    return {
-      accessToken: body.data.accessToken,
-      setCookies,
-      message: body.message || 'Success',
-    };
-  },
-};
